Allow reusing a client instance
Change the implementation so the client needs to be initialized only once and can be reused to make requests concurrently.

For that purpose, the V4 signature logic (ahem blindly copy/pasted from AWS code) don’t use instance variables to store the date when the client was initialized. Instead, the date is taken when a request is issued.

Also slightly refactored the code to follow some Java best practices. Most of them were suggested by an IDE though.